% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_jumps.R
\name{find_jumps}
\alias{find_jumps}
\title{Find dispersal jumps}
\usage{
find_jumps(grid_data = grid_data, potJumps = potJumps, gap_size = 15)
}
\arguments{
\item{grid_data}{A data frame of unique points to be processed, i.e., the initial grid_data table}

\item{potJumps}{A data frame of potential jumps to be refined, resulting from find_thresholds()}

\item{gap_size}{Distance between the invasion front and the positive point
necessary for it to be considered a jump, in kilometers (default: 15)}
}
\value{
Two data frames: \code{Jumps} containing the list of jumps, and \code{potSecDiff} containing
potential secondary diffusion to be processed using find_secDiff()
}
\description{
Find dispersal jumps after having applied find_thresholds(). Prune points selected due to sector limits and secondary diffusion.
}
\examples{
\dontrun{
jumps <- find_jumps(dataset)
}
}
