% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_jumps_wrapper.R
\name{find_jumps_wrapper}
\alias{find_jumps_wrapper}
\title{Find dispersal jumps}
\usage{
find_jumps_wrapper(
  dataset = grid_data,
  nb_sectors = 12,
  centroid = c(-75.67534, 40.41524),
  gap_size = 15,
  negatives = T
)
}
\arguments{
\item{dataset}{A data frame of unique points to be processed, ideally the dataset_unique resulting from find_thresholds()}

\item{nb_sectors}{Number of sectors in which the invasion range is divided}

\item{centroid}{long,lat coordinates of the introduction point}

\item{gap_size}{Distance between the invasion front and the positive point
necessary for it to be considered a jump, in kilometers (default: 15)}

\item{negatives}{Should negative surveys be considered? (default: TRUE)}
}
\value{
Three data frames: \code{Jumps} containing all jumps, \code{Dist} containing the 
limits of the diffusive spread, and \code{secDiff} containing secondary diffusion cells
}
\description{
Find dispersal jumps after having applied find_thresholds(). Prune points selected due to sector limits and secondary diffusion.
}
\examples{
\dontrun{
jumps <- find_jumps(dataset)
}
}
