% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_secDiff.R
\name{find_secDiff}
\alias{find_secDiff}
\title{Find secondary diffusion}
\usage{
find_secDiff(
  potDiffusion = potDiffusion,
  Jumps = Jumps,
  diffusers = diffusers,
  Dist = preDist,
  gap_size = 15
)
}
\arguments{
\item{potDiffusion}{object exported from find_jumps}

\item{Jumps}{object exported from find_jumps}

\item{diffusers}{object exported from find_jumps}

\item{Dist}{object exported from find_jumps}

\item{gap_size}{Distance necessary for a point to be considered a jump, 
in kilometers (default: 15)}
}
\value{
A data frame \code{secDiff} containing the secondary diffusion points, a data frame \code{Dist} containing the final thresholds of the diffusive spread.
}
\description{
Find secondary diffusion points after having applied find_jumps()
}
\examples{
\dontrun{
secDiff <- find_secDiff()
}
}
