% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_thresholds.R
\name{find_thresholds}
\alias{find_thresholds}
\title{Find thresholds of diffusive spread}
\usage{
find_thresholds(dataset, gap_size = 15, negatives = TRUE)
}
\arguments{
\item{dataset}{A dataset to be processed, output of attribute_sectors()}

\item{gap_size}{Distance between the invasion front and a positive point
necessary for it to be considered a jump, in kilometers (default: 15)}

\item{negatives}{Should the function verify that negative surveys were found 
in the spatial discontinuities? (default: TRUE)}
}
\value{
Two tables: one table \code{Thresholds} containing the threshold per year and sector,
one table \code{potJumps} containing the list of potential jumps for find_jumps()
}
\description{
Find thresholds of diffusive spread and get a list of potential jumps. 
This function identifies spatial discontinuities in the distribution of 
species’ occurrences between the putative limit of the continuous, diffusive 
spread (i.e., the invasion front) and outlying occurrences.
This function must be applied after attribute_sectors() and before find_jumps().
}
\examples{
\dontrun{
thresholds <- find_thresholds(dataset)
}
}
